+++ testsuites/TestsuiteGT/performance-test/org/emoflon/ibex/gt/performance/SheRememberedCaterpillarsPerformanceTest2.java
@@ -67,7 +67,7 @@ public class SheRememberedCaterpillarsPerformanceTest2 extends GTPerformanceTest
 		
 		// Execute query.
 		long queryStart = System.nanoTime();
-		int count = api.findStandalonePlatform().countMatches() + api.findEmptyExit().countMatches();
+		long count = api.findStandalonePlatform().countMatches() + api.findEmptyExit().countMatches();
 		if (count != modelSize) {
 			System.out.println("ERROR");
 		}
